Refactor to Arrow code
Notice that the word 'refactor' here is used entirely in the meaning of
'rearranging the code without changing the behaviour of the system'.
Here, 'not changing the behaviour' means that all tests still pass. In
fact, this change exacerbates the code.

This is most notable in the over-confident exception message in the
bottom of the method. There are definitely edge cases not handled by
this variation on the code, but 'the developer' left a comment because
(s)he couldn't deal with all the possible variations.

Another problem is the duplication of code, as demonstrated by the three
(why three?) duplications of the `Connect` and `Update` code.