Introduce result return object for lookup method
Sure enough, the code reviewer wasn't pleased with the use of an `out`
parameter, which 'didn't feel sufficiently object-oriented'. Our poor
implementer now tries to introduce a 'result' object, but as the call
sites amply illustrate, this doesn't seem to make the API easier to use.

To add spite to injury, one now has to initialise the User objects to
null, because clearly, they may never get assigned, and the compiler is
actually clever enough to detect that.